pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "wlan-localization"
version = "2.0.0"
description = "Machine Learning approach to Wi-Fi fingerprint-based indoor localization"
readme = "README.md"
requires-python = ">=3.11"
license = {text = "MIT"}
authors = [
    {name = "Sharan Naribole", email = "sharan.naribole@example.com"}
]
keywords = ["machine-learning", "indoor-localization", "wifi-fingerprinting", "spatial-analysis"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]

dependencies = [
    "numpy>=1.26.0,<2.0.0",
    "pandas>=2.2.0",
    "scikit-learn>=1.4.0",
    "scipy>=1.12.0",
    "matplotlib>=3.8.0",
    "seaborn>=0.13.0",
    "imbalanced-learn>=0.12.0",
    "xgboost>=2.0.0",
    "pyyaml>=6.0",
    "loguru>=0.7.0",
    "click>=8.1.0",
    "joblib>=1.3.0",
    "mlflow>=2.10.0",
    "optuna>=3.5.0",
    "pydantic>=2.5.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-cov>=4.1.0",
    "pytest-xdist>=3.5.0",
    "hypothesis>=6.98.0",
    "black>=24.1.0",
    "isort>=5.13.0",
    "flake8>=7.0.0",
    "mypy>=1.8.0",
    "pre-commit>=3.6.0",
    "ipython>=8.20.0",
    "jupyter>=1.0.0",
    "notebook>=7.0.0",
]

docs = [
    "sphinx>=7.2.0",
    "sphinx-rtd-theme>=2.0.0",
    "myst-parser>=2.0.0",
]

all = [
    "wlan-localization[dev,docs]",
]

[project.scripts]
wlan-train = "wlan_localization.cli.train:main"
wlan-evaluate = "wlan_localization.cli.evaluate:main"
wlan-predict = "wlan_localization.cli.predict:main"

[project.urls]
Homepage = "https://github.com/yourusername/wlan_localization"
Documentation = "https://github.com/yourusername/wlan_localization/blob/master/README.md"
Repository = "https://github.com/yourusername/wlan_localization"
Issues = "https://github.com/yourusername/wlan_localization/issues"

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-data]
wlan_localization = ["py.typed"]

[tool.black]
line-length = 100
target-version = ['py311']
include = '\.pyi?$'
extend-exclude = '''
/(
  # directories
  \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | build
  | dist
)/
'''

[tool.isort]
profile = "black"
line_length = 100
multi_line_output = 3
include_trailing_comma = true
force_grid_wrap = 0
use_parentheses = true
ensure_newline_before_comments = true

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
strict_equality = true

[[tool.mypy.overrides]]
module = [
    "sklearn.*",
    "scipy.*",
    "matplotlib.*",
    "seaborn.*",
    "imblearn.*",
    "xgboost.*",
    "mlflow.*",
    "optuna.*",
]
ignore_missing_imports = true

[tool.pytest.ini_options]
minversion = "8.0"
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = [
    "--verbose",
    "--cov=wlan_localization",
    "--cov-report=html",
    "--cov-report=term-missing",
    "--cov-report=xml",
    "--cov-fail-under=80",
]

[tool.coverage.run]
source = ["src"]
omit = [
    "*/tests/*",
    "*/test_*.py",
]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise AssertionError",
    "raise NotImplementedError",
    "if __name__ == .__main__.:",
    "if TYPE_CHECKING:",
    "class .*\\bProtocol\\):",
    "@(abc\\.)?abstractmethod",
]
